requirements.txt
streamlit==1.24.0
tensorflow-cpu==2.10.0
Pillow==9.3.0
numpy==1.23.5
